+++ src/domain/application/application.resolver.ts
@@ -1,11 +1,11 @@
 import { Resolver, Query, Args, Mutation } from '@nestjs/graphql';
 import { Application } from '@domain/application/application.entity';
 import { ApplicationService } from '@domain/application/application.service';
-import { Roles } from '@utils/authorisation/roles.decorator';
+import { Roles } from '@utils/authorization/roles.decorator';
 import { UseGuards } from '@nestjs/common';
-import { GqlAuthGuard } from '@utils/authorisation/graphql.guard';
+import { GqlAuthGuard } from '@utils/authorization/graphql.guard';
 import { Profiling } from '@utils/logging/logging.profiling.decorator';
-import { AuthorisationRoles } from '@utils/authorisation/authorisation.roles';
+import { AuthorisationRoles } from '@utils/authorization/authorization.roles';
 
 @Resolver(() => Application)
 export class ApplicationResolver {
